+++ benchmarks/small-manual.ts
@@ -1,3 +1,7 @@
+/**
+ * *-manual.ts is where end user specifiy which fields should be sanitized manually
+ **/
+
 import { t } from 'elysia'
 import { benchmark } from './utils'
 
@@ -5,7 +9,9 @@ benchmark(
 	t.Object({
 		id: t.Number(),
 		name: t.String(),
-		bio: t.String(),
+		bio: t.String({
+			sanitize: true
+		}),
 		metadata: t.Object({
 			alias: t.String(),
 			country: t.String()
@@ -14,10 +20,13 @@ benchmark(
 	{
 		id: 1,
 		name: 'SaltyAom',
-		bio: 'I like train',
+		bio: 'I like train\n',
 		metadata: {
 			alias: 'SaltyAom',
 			country: 'Thailand'
 		}
+	},
+	{
+		sanitize: 'manual'
 	}
 )